Update emg_predictor.py
Improve shared memory tracking for online streamer. Previously we didn't have a way of knowing when the shared memory was updated. Similar to other parts of the code, introducing a _count tag should help alleviate that issue. Note that it appears that `classifier_smm_writes` was being used for something similar, but given that it was initialized as process-local variable, this won't be reflected back to the parent if not running in blocking mode.